cleaned up a bit

+++ libft/src/mandatory/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: lverdoes <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/30 14:17:12 by lverdoes      #+#    #+#                 */
-/*   Updated: 2021/02/26 15:05:59 by lverdoes      ########   odam.nl         */
+/*   Updated: 2021/07/03 10:28:05 by lverdoes      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_atoi(const char *str)
 {
-	unsigned long long	result;
-	int					neg;
-	size_t				i;
+	size_t	result;
+	int		neg;
+	size_t	i;
 
 	result = 0;
 	i = 0;
